requirements.txt
attrs==17.4.0
beautifulsoup4==4.7.1
Flask==0.10.1
lxml==4.1.1
opencv-python==4.2.0.34
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.11
py==1.5.2
pytest==3.4.1
python-docx==0.8.10
python-pptx==0.6.18
Pillow==5.4.1
lxml==4.5.0
requests==2.21.0
six==1.11.0
XlsxWriter==1.2.8